Modified SQL install script (button_id type)

+++ app/code/local/Accolade/Bttn/sql/mysql4-install-0.1.0.php
@@ -10,7 +10,7 @@ DROP TABLE IF EXISTS {$this->getTable('accolade_bttn')};
 CREATE TABLE {$this->getTable('accolade_bttn')} (
     `entity_id` int(11) unsigned NOT NULL AUTO_INCREMENT,  
     `customer_id` int(16) unsigned NOT NULL,  
-    `button_id` int(16) unsigned NOT NULL,  
+    `button_id` bigint(16) unsigned NOT NULL,  
     `shipping_method` varchar(40) NOT NULL,  
     `payment_method` varchar(40) NOT NULL,  
     PRIMARY KEY (`entity_id`)
